+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.7.0] - 2025-11-19
+### Added
+- Call to mercado pago woocommerce scripts
+### Changed
+- Ensure that card security code will be validated before form submit on fast payment flow.
+- Only send pre-load fast payment metric when flow is supported
+- Use new layout to show fast payment flow payment methods
+- Improvement error message treatment on checkout custom
+### Fixed
+- Fix checkout custom layout
+
 ## [8.6.1] - 2025-11-06
 ### Added
 - Added card form click metric to Checkout API funnel tracking